feat: solve missing dependencies, circular deps
