+++ chat-challenge/app.json
@@ -36,7 +36,8 @@
           "microphonePermission": "Allow $(PRODUCT_NAME) to access your microphone",
           "recordAudioAndroid": true
         }
-      ]
+      ],
+      "expo-font"
     ],
     "owner": "thiago.dasilva"
   }
