+++ chat-challenge/package.json
@@ -16,21 +16,23 @@
     "@tamagui/babel-plugin": "^1.89.17",
     "@tamagui/config": "1.89.17",
     "@tamagui/core": "1.89.17",
+    "@tamagui/font-inter": "^1.116.9",
     "@tamagui/lucide-icons": "1.89.17",
     "expo": "~51.0.28",
+    "expo-camera": "~15.0.16",
+    "expo-font": "~12.0.10",
     "expo-status-bar": "~1.12.1",
     "openai": "^4.68.4",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-native": "0.74.5",
+    "react-native-gesture-handler": "~2.16.1",
     "react-native-get-random-values": "^1.11.0",
     "react-native-gifted-chat": "^2.6.4",
     "react-native-reanimated": "3.10.1",
     "react-native-safe-area-context": "^4.12.0",
     "react-native-svg": "^15.8.0",
-    "tamagui": "1.89.17",
-    "expo-camera": "~15.0.16",
-    "react-native-gesture-handler": "~2.16.1"
+    "tamagui": "1.89.17"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
