+++ chat-challenge/src/ui/organisms/screens/CameraScreen/index.ts
@@ -1,2 +1,1 @@
 export * from './CameraScreen'
-export * from './CameraStore'
